Re-organize the tests
