+++ tests/test_mongoengine/test_order_by.py
@@ -1,8 +1,6 @@
 import pytest
 from pydantic import ValidationError
 
-from fastapi_filter.contrib.mongoengine import Filter
-
 
 @pytest.mark.parametrize(
     "order_by,assert_function",
@@ -26,15 +24,9 @@ from fastapi_filter.contrib.mongoengine import Filter
         ],
     ],
 )
-def test_basic_order_by(User, users, order_by, assert_function):
-    class UserFilter(Filter):
-        class Constants(Filter.Constants):
-            model = User
-
-        order_by: list[str] | None
-
+def test_basic_order_by(User, users, UserFilterOrderBy, order_by, assert_function):
     query = User.objects().all()
-    query = UserFilter(order_by=order_by).sort(query)
+    query = UserFilterOrderBy(order_by=order_by).sort(query)
     previous_user = None
     for user in query:
         if not previous_user:
@@ -44,15 +36,48 @@ def test_basic_order_by(User, users, order_by, assert_function):
         previous_user = user
 
 
-def test_order_by_with_default(User, users):
-    class UserFilter(Filter):
-        class Constants(Filter.Constants):
-            model = User
+@pytest.mark.parametrize(
+    "order_by,assert_function",
+    [
+        [None, lambda previous_user, user: True],
+        [
+            "name",
+            lambda previous_user, user: previous_user["name"] <= user["name"]
+            if previous_user["name"] and user["name"]
+            else True,
+        ],
+        [
+            "-created_at",
+            lambda previous_user, user: previous_user["created_at"] >= user["created_at"],
+        ],
+        [
+            "age,-name",
+            lambda previous_user, user: (previous_user["age"] < user["age"])
+            or (
+                previous_user["age"] == user["age"]
+                and (previous_user["name"] <= user["name"] if previous_user["name"] and user["name"] else True)
+            ),
+        ],
+    ],
+)
+@pytest.mark.asyncio
+async def test_api_basic_order_by(test_client, users, order_by, assert_function):
+    endpoint = "/users"
+    if order_by is not None:
+        endpoint = f"{endpoint}?order_by={order_by}"
+    response = await test_client.get(endpoint)
+    previous_user = None
+    for user in response.json():
+        if not previous_user:
+            previous_user = user
+            continue
+        assert assert_function(previous_user, user)
+        previous_user = user
 
-        order_by: list[str] = ["age"]
 
+def test_order_by_with_default(User, users, UserFilterOrderByWithDefault):
     query = User.objects().all()
-    query = UserFilter().sort(query)
+    query = UserFilterOrderByWithDefault().sort(query)
     previous_user = None
     for user in query:
         if not previous_user:
@@ -62,23 +87,13 @@ def test_order_by_with_default(User, users):
         previous_user = user
 
 
-def test_invalid_order_by(User, users):
-    class UserFilter(Filter):
-        class Constants(Filter.Constants):
-            model = User
-
-        order_by: list[str] | None
-
+def test_invalid_order_by(User, users, UserFilterOrderBy):
     with pytest.raises(ValidationError):
-        UserFilter(order_by="invalid")
+        UserFilterOrderBy(order_by="invalid")
 
 
-def test_missing_order_by_field(User):
-    class UserFilter(Filter):
-        class Constants(Filter.Constants):
-            model = User
-
+def test_missing_order_by_field(User, UserFilterNoOrderBy):
     query = User.objects().all()
 
     with pytest.raises(AttributeError):
-        UserFilter().sort(query)
+        UserFilterNoOrderBy().sort(query)
